+++ feature/projects/src/main/AndroidManifest.xml
@@ -8,7 +8,7 @@
             android:theme="@style/Theme.AppCompat.Light.NoActionBar"
             android:exported="false" />
         <activity
-            android:name="com.zucchini.projects.ProjectDetailActivity"
+            android:name="com.zucchini.projects.projects.ProjectDetailActivity"
             android:exported="false"
             android:screenOrientation="portrait"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar" />
@@ -18,7 +18,7 @@
             android:screenOrientation="portrait"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar" />
         <activity
-            android:name="com.zucchini.projects.DevDetailActivity"
+            android:name="com.zucchini.projects.developer.DevDetailActivity"
             android:exported="false"
             android:screenOrientation="portrait"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar" />
