+++ feature/projects/src/main/java/com/zucchini/projects/mypage/MypageFragment.kt
@@ -1,4 +1,4 @@
-package com.zucchini.projects
+package com.zucchini.projects.mypage
 
 import android.content.Intent
 import android.net.Uri
