+++ feature/projects/src/main/res/layout/activity_project_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.zucchini.projects.ProjectDetailActivity">
+    tools:context="com.zucchini.projects.projects.ProjectDetailActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
         android:layout_width="match_parent"
