+++ feature/projects/src/main/res/layout/fragment_dev_info.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.zucchini.projects.DevInfoFragment">
+    tools:context="com.zucchini.projects.developer.DevInfoFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/cl_top_app_bar"
